requirements.txt
fastapi[all]
strawberry-graphql[fastapi]
uvicorn[standar]
sqlalchemy
requests
PyJWT
python-dotenv
email-validator